+++ tests/test_datawriter.py
@@ -5,8 +5,8 @@
 
 # from tensorpack import *
 from tensorpack.dataflow.base import DataFlow
-from tensorpack.dataflow.dftools import LMDBDataWriter, TFRecordDataWriter, NumpyDataWriter, HDF5DataWriter
-from tensorpack.dataflow.format import LMDBDataReader, TFRecordDataReader, NumpyDataReader, HDF5DataReader
+from tensorpack.dataflow.serializer import LMDBDataSerializer, TFRecordDataSerializer
+from tensorpack.dataflow.serializer import NumpyDataSerializer, HDF5DataSerializer
 import unittest
 import os
 import numpy as np
@@ -21,6 +21,7 @@ def delete_file_if_exists(fn):
 
 class SeededFakeDataFlow(DataFlow):
     """docstring for SeededFakeDataFlow"""
+
     def __init__(self, seed=42, size=32):
         super(SeededFakeDataFlow, self).__init__()
         self.seed = seed
@@ -44,15 +45,13 @@ class SeededFakeDataFlow(DataFlow):
 
 class SerializerTest(unittest.TestCase):
 
-    def run_write_read_test(self, file, writer, w_args, w_kwargs, reader, r_args, r_kwargs, error_msg):
+    def run_write_read_test(self, file, serializer, w_args, w_kwargs, r_args, r_kwargs, error_msg):
         try:
             delete_file_if_exists(file)
 
             ds_expected = SeededFakeDataFlow()
-            serializer = writer(ds_expected, file, *w_args, **w_kwargs)
-            serializer.serialize()
-
-            ds_actual = reader(file, *r_args, **r_kwargs)
+            serializer(file).save(ds_expected, *w_args, **w_kwargs)
+            ds_actual = serializer(file).load(*r_args, **r_kwargs)
 
             ds_actual.reset_state()
             ds_expected.reset_state()
@@ -64,20 +63,28 @@ class SerializerTest(unittest.TestCase):
             print(error_msg)
 
     def test_lmdb(self):
-        self.run_write_read_test('test.lmdb', LMDBDataWriter, {}, {}, LMDBDataReader, {}, {'shuffle': False},
+        self.run_write_read_test('test.lmdb', LMDBDataSerializer,
+                                 {}, {'overwrite': False},
+                                 {}, {'shuffle': False},
                                  'skip test_lmdb, no lmdb available')
 
     def test_tfrecord(self):
-        self.run_write_read_test('test.tfrecord', TFRecordDataWriter, {}, {}, TFRecordDataReader, {}, {'size': 32},
+        self.run_write_read_test('test.tfrecord', TFRecordDataSerializer,
+                                 {}, {'overwrite': False},
+                                 {}, {'size': 32},
                                  'skip test_tfrecord, no tensorflow available')
 
     def test_numpy(self):
-        self.run_write_read_test('test.npz', NumpyDataWriter, {}, {}, NumpyDataReader, {}, {'shuffle': False},
+        self.run_write_read_test('test.npz', NumpyDataSerializer,
+                                 {}, {'overwrite': False},
+                                 {}, {'shuffle': False},
                                  'skip test_numpy, no numpy available')
 
     def test_hdf5(self):
         args = [['label', 'image']]
-        self.run_write_read_test('test.npz', HDF5DataWriter, args, {}, HDF5DataReader, args, {'shuffle': False},
+        self.run_write_read_test('test.npz', HDF5DataSerializer,
+                                 args, {'overwrite': False},
+                                 args, {'shuffle': False},
                                  'skip test_hdf5, no h5py available')
